Remove hardcoded accessor for time in prep for change in basicShapes
